requirements.txt
flask
fitz  # PyMuPDF
sentence-transformers
faiss-cpu
numpy
scikit-learn
nltk
tqdm
